Update MainUIPhoto.swift
Removed commented out bad code (the code that caused incorrect resolutions to be filtered and stored).

Fixed bug that caused original images to be safed as .tif files by using the same code that saved filtered images. This code was moved to a new function so both paths can use it.