Update AppDelegate.swift
Added comments to major state change events. Added global variables for running with a debugger, previous instance crashed, and the crashed filter name. Added code to determine if running via a debugger. Added code to detect if the previous instance crashed and to reset filters if that was the case.