Update FilterSettingUIBase.swift
Added parameter to Initialize that tells the class whether it is the base class for the root filter settings UI or a child controller of the root. If for the root, fave and stars appear in the toolbar, otherwise not. Back button also behaves differently (but as the user expects).

Removed the table cell view that holds the fave and star controls. Added fave and star controls to the toolbar at the bottom and removed the Camera Home button.