Update FilterSettingUIBase.swift
Removed commented out code that added ratings and fave controls to the Settings UI table view as a separate table view cell. Removed HandleCameraHomeBButton function (which was empty anyway). Added check in all tableView overrides to ensure we manipulate the proper tableview. Found pesky bug that allowed the user to tap the sample view to bring up a menu - commented out code that enabled that behavior.